refactor legacy config
