+++ src/app/legacy/config/index.ts
@@ -3,7 +3,6 @@ import {
   IndexDebtIssuanceModuleV2Address,
   IndexDebtIssuanceModuleV2Address_v2,
 } from '@indexcoop/flash-mint-sdk'
-import { getTokenByChainAndSymbol } from '@indexcoop/tokenlists'
 
 import { LegacyToken } from '@/app/legacy/types'
 import {
@@ -14,6 +13,7 @@ import {
   LeveragedRethStakingYield,
 } from '@/constants/tokens'
 
+import { BedIndex, GitcoinStakedETHIndex, ic21 } from './tokens/mainnet'
 import {
   Bitcoin2xFlexibleLeverageIndexPolygon,
   ETH2xFlexibleLeverageIndexPolygon,
@@ -21,15 +21,11 @@ import {
   InverseETHFlexibleLeverageIndexPolygon,
   InverseMATICFlexibleLeverageIndexPolygon,
   Matic2xFlexibleLeverageIndexPolygon,
-} from './polygon'
+} from './tokens/polygon'
 
 const DebtIssuanceModuleAddress = '0x39F024d621367C044BacE2bf0Fb15Fb3612eCB92'
 const DebtIssuanceModuleV2PolygonAddress =
   '0xf2dC2f456b98Af9A6bEEa072AF152a7b0EaA40C9'
-
-const BedIndex = getTokenByChainAndSymbol(1, 'BED')
-const GitcoinStakedETHIndex = getTokenByChainAndSymbol(1, 'gtcETH')
-const ic21 = getTokenByChainAndSymbol(1, 'ic21')
 
 export const Issuance = {
   [BedIndex.symbol]: IndexDebtIssuanceModuleV2Address_v2,
