+++ src/app/legacy/config/tokens/mainnet.ts
@@ -1,0 +1,5 @@
+import { getTokenByChainAndSymbol } from '@indexcoop/tokenlists'
+
+export const BedIndex = getTokenByChainAndSymbol(1, 'BED')
+export const GitcoinStakedETHIndex = getTokenByChainAndSymbol(1, 'gtcETH')
+export const ic21 = getTokenByChainAndSymbol(1, 'ic21')
